Added order.turnover method, basic regex email validator and an invalid email exception.

+++ src/main/java/com/ivelinnikolov/ProjectSAP/controllers/ProductController.java
@@ -4,8 +4,10 @@ import com.ivelinnikolov.ProjectSAP.models.Product;
 import com.ivelinnikolov.ProjectSAP.services.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
 import java.util.List;
 
@@ -25,21 +27,21 @@ public class ProductController
         return productService.listAll();
     }
 
-    @GetMapping("/product-amount/{id}")
+    @GetMapping("/amount/{id}")
     @ResponseStatus(HttpStatus.OK)
     public int showProductAmount(@PathVariable(name = "id") int id)
     {
         return productService.listProductAmount(id);
     }
 
-    @GetMapping("/list-by-id/{id}")
+    @GetMapping("/{id}")
     @ResponseStatus(HttpStatus.OK)
     public Product getProduct(@PathVariable(name = "id") int id)
     {
         return  productService.getProductById(id);
     }
 
-    @GetMapping("/list-all-discounted-items")
+    @GetMapping("/discount")
     @ResponseStatus(HttpStatus.OK)
     public List<Product> getAllDiscountedProducts()
     {
@@ -48,36 +50,73 @@ public class ProductController
 
     //-------------------------------------------- PUT REQUESTS --------------------------------------------
 
-    @PutMapping("/update-product-price-{id}/{newPrice}/{newMinPrice}")
+    @PutMapping("/update-{id}/{newPrice}/{newMinPrice}")
     @ResponseStatus(HttpStatus.OK)
-    public Product updatePricesById(@PathVariable(name = "id") int id, @PathVariable(name = "newPrice") double newPrice,
-    @PathVariable(name = "newMinPrice") double newMinPrice)
+    public ResponseEntity<?> updatePricesById(@PathVariable(name = "id") int id, @PathVariable(name = "newPrice") double newPrice,
+    @PathVariable(name = "newMinPrice") double newMinPrice, HttpSession session)
     {
-        return productService.updatePricesByID(id, newPrice, newMinPrice);
+        if (session.getAttribute("userId") == null)
+        {
+            return ResponseEntity.status(401).body("Unauthorized!");
+        }
+        else if (!session.getAttribute("userRole").equals(2)) // 2 - employee
+        {
+            return ResponseEntity.status(403).body("Request unavailable for this account.");
+        }
+
+        return ResponseEntity.ok(productService.updatePricesByID(id, newPrice, newMinPrice));
     }
 
-    @PutMapping("/update-product-{id}")
+    @PutMapping("/update-{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Product updateProductById(@PathVariable(name = "id") int id, @Valid @RequestBody Product updatedProduct)
+    public ResponseEntity<?> updateProductById(@PathVariable(name = "id") int id, @Valid @RequestBody Product updatedProduct, HttpSession session)
     {
-        return productService.updateProductById(id, updatedProduct);
+        if (session.getAttribute("userId") == null)
+        {
+            return ResponseEntity.status(401).body("Unauthorized!");
+        }
+        else if (!session.getAttribute("userRole").equals(2)) // 2 - employee
+        {
+            return ResponseEntity.status(403).body("Request unavailable for this account.");
+        }
+
+        return ResponseEntity.ok(productService.updateProductById(id, updatedProduct));
     }
 
     //-------------------------------------------- POST REQUESTS --------------------------------------------
 
     @PostMapping("/add")
     @ResponseStatus(HttpStatus.CREATED)
-    public Product createProduct(@Valid @RequestBody Product product)
+    public ResponseEntity<?> createProduct(@Valid @RequestBody Product product, HttpSession session)
     {
-        return productService.createNewProduct(product);
+        if (session.getAttribute("userId") == null)
+        {
+            return ResponseEntity.status(401).body("Unauthorized!");
+        }
+        else if (!session.getAttribute("userRole").equals(2)) // 2 - employee
+        {
+            return ResponseEntity.status(403).body("Request unavailable for this account.");
+        }
+
+        return ResponseEntity.ok(productService.createNewProduct(product));
     }
 
     //-------------------------------------------- DELETE REQUESTS --------------------------------------------
 
-    @DeleteMapping("/remove-product-by-id-{id}")
+    @DeleteMapping("/remove-{id}")
     @ResponseStatus(HttpStatus.OK)
-    public void deleteProductByID(@PathVariable(value = "id") Integer accID)
+    public ResponseEntity<?> deleteProductByID(@PathVariable(value = "id") Integer accID, HttpSession session)
     {
+        if (session.getAttribute("userId") == null)
+        {
+            return ResponseEntity.status(401).body("Unauthorized!");
+        }
+        else if (!session.getAttribute("userRole").equals(2)) // 2 - employee
+        {
+            return ResponseEntity.status(403).body("Request unavailable for this account.");
+        }
+
         productService.deleteByID(accID);
+        return ResponseEntity.noContent().build();
     }
 }
